pyproject.toml
[tool.poetry]
name = "pos-ml"
version = "0.1.0"
description = ""
authors = ["Leandro-barreto <leandro.lhbb@gmail.com>"]
readme = "README.md"
packages = [{include = "pos_ml"}]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^2.2.3"
jupyter = "^1.1.1"
pytest = "^8.3.3"
wheel = "^0.44.0"
twine = "^5.1.1"
flask = "^3.0.3"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
flask-sqlalchemy = "^3.1.1"
investimentpy = "^1.0.3"
matplotlib = "^3.9.2"
beautifulsoup4 = "^4.12.3"
pyjwt = "^2.10.1"
pillow = "^11.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
